+++ nic.go
@@ -5,10 +5,10 @@ import (
 )
 
 const (
-	version   = "0.2.1"
-	userAgent = "golang-nic/0.2.1"
-	author    = "Iv4n"
-	copyright = "Copyright 2019 Iv4n"
+	version   = "0.3"
+	userAgent = "golang-nic/0.3"
+	author    = "iv4n"
+	copyright = "Copyright 2019 iv4n"
 )
 
 var (